+++ public/app.js
@@ -180,15 +180,18 @@ var runThis = world => {
         _manifest.default.scene.player.energy = {}; // test 
 
         addEventListener('hit.keyDown', e => {
-          console.log('Bring to the top level', e.detail.keyCode); // dont mess in events 
+          // console.log('Bring to the top level', e.detail.keyCode);
+          // dont mess in events 
           // better in high level to be respoved
-
           if (e.detail.keyCode == 32) {
             setTimeout(() => {
               _manifest.default.scene.playerCollisonBox.physics.currentBody.force.set(0, 0, 111);
             }, 250);
           } else if (e.detail.keyCode == 87) {
-            _manifest.default.scene.playerCollisonBox.physics.currentBody.force.set(0, 10, 0);
+            // Good place for blocking volume
+            // App.scene.playerCollisonBox.physics.currentBody.force.set(0,10,0)
+            setTimeout(() => {// App.scene.playerCollisonBox.physics.currentBody.force.set(0,100,0)
+            }, 100);
           }
         });
         var playerUpdater = {
@@ -207,35 +210,31 @@ var runThis = world => {
             }
 
             if (matrixEngine.Events.camera.virtualJumpActive == true) {
-              console.log('?????'); // invert logic
+              // invert logic
               // Scene object set
-
               _manifest.default.scene[objName].rotation.rotateX(-detPitch);
 
               var detPitchPos = matrixEngine.Events.camera.pitch;
-              if (detPitchPos > 4) detPitchPos = 4; // App.scene.playerCollisonBox.physics.currentBody.force.set(0, 0, 1);
+              if (detPitchPos > 4) detPitchPos = 4;
 
-              _manifest.default.scene[objName].position.setPosition(_manifest.default.scene.playerCollisonBox.physics.currentBody.position.x, _manifest.default.scene.playerCollisonBox.physics.currentBody.position.z, _manifest.default.scene.playerCollisonBox.physics.currentBody.position.y); // Cannonjs object set
-              // Switched  Z - Y
+              _manifest.default.scene.playerCollisonBox.physics.currentBody.force.set(0, 0, 70);
+
+              _manifest.default.scene[objName].position.setPosition(_manifest.default.scene.playerCollisonBox.physics.currentBody.position.x, _manifest.default.scene.playerCollisonBox.physics.currentBody.position.z, _manifest.default.scene.playerCollisonBox.physics.currentBody.position.y); // Cannonjs object set / Switched  Z - Y
 
 
               matrixEngine.Events.camera.xPos = _manifest.default.scene.playerCollisonBox.physics.currentBody.position.x;
               matrixEngine.Events.camera.zPos = _manifest.default.scene.playerCollisonBox.physics.currentBody.position.y;
-              matrixEngine.Events.camera.yPos = _manifest.default.scene.playerCollisonBox.physics.currentBody.position.z; // App.scene.playerCollisonBox.
-              //   physics.currentBody.velocity.set(0, 0, 0);
+              matrixEngine.Events.camera.yPos = _manifest.default.scene.playerCollisonBox.physics.currentBody.position.z;
 
               _manifest.default.scene.playerCollisonBox.physics.currentBody.angularVelocity.set(0, 0, 0);
 
               setTimeout(() => {
-                matrixEngine.Events.camera.virtualJumpY = 2;
                 matrixEngine.Events.camera.virtualJumpActive = false;
               }, 1350);
             } else {
               // Tamo tu iznad duge nebo zri...
               // Cannonjs object set
               // Switched  Z - Y
-              //matrixEngine.Events.camera.xPos = App.scene.playerCollisonBox.physics.currentBody.position.x;
-              // matrixEngine.Events.camera.zPos = App.scene.playerCollisonBox.physics.currentBody.position.y;
               matrixEngine.Events.camera.yPos = _manifest.default.scene.playerCollisonBox.physics.currentBody.position.z; // Scene object set
 
               _manifest.default.scene[objName].rotation.rotateX(-detPitch);
@@ -248,9 +247,7 @@ var runThis = world => {
               // Switched  Z - Y
 
 
-              _manifest.default.scene.playerCollisonBox.physics.currentBody.position.set(matrixEngine.Events.camera.xPos, matrixEngine.Events.camera.zPos, matrixEngine.Events.camera.yPos); //                 App.scene.playerCollisonBox.physics.currentBody.position.y);
-              // App.scene.playerCollisonBox.physics.currentBody.velocity.set(0, 0, 0);
-              // App.scene.playerCollisonBox.physics.currentBody.angularVelocity.set(0, 0, 0);
+              _manifest.default.scene.playerCollisonBox.physics.currentBody.position.set(matrixEngine.Events.camera.xPos, matrixEngine.Events.camera.zPos, matrixEngine.Events.camera.yPos); // App.scene.playerCollisonBox.physics.currentBody.position.y)
 
             }
           }
@@ -399,15 +396,14 @@ var runThis = world => {
         }, 1000 * j);
       }));
     }
-  };
+  }; // objGenerator(15);
 
-  objGenerator(15);
+
   createObjSequence('player');
   Promise.all(promiseAllGenerated).then(what => {
     console.info(`Runtime wait for some generetion of scene objects,
                   then swap scene array index for target -> 
-                  must be manual setup for now!`, what);
-    swap(5, 19, matrixEngine.matrixWorld.world.contentList);
+                  must be manual setup for now!`, what); // swap(5, 19, matrixEngine.matrixWorld.world.contentList);
   }); // Add ground for physics bodies.
 
   var tex = {
@@ -470,6 +466,22 @@ var runThis = world => {
   _manifest.default.scene.energy.geometry.setScaleByX(0.35);
 
   _manifest.default.scene.energy.geometry.setScaleByY(0.1);
+
+  world.Add("cubeLightTex", 1, "FLOOR2", tex);
+  var b2 = new CANNON.Body({
+    mass: 0,
+    linearDamping: 0.01,
+    position: new CANNON.Vec3(1, -14.5, -1),
+    shape: new CANNON.Box(new CANNON.Vec3(3, 1, 1))
+  });
+  physics.world.addBody(b2);
+
+  _manifest.default.scene['FLOOR2'].position.setPosition(1, -1, -14.5);
+
+  _manifest.default.scene['FLOOR2'].geometry.setScaleByX(3);
+
+  _manifest.default.scene['FLOOR2'].physics.currentBody = b2;
+  _manifest.default.scene['FLOOR2'].physics.enabled = true;
 };
 
 exports.runThis = runThis;
@@ -2034,9 +2046,10 @@ camera.setCamera = function (object) {
     if (camera.fly == true) {
       // Fly regime
       camera.yPos += Math.sin(degToRad(camera.pitch)) * camera.speed;
-    } else {
-      // usually for fpshooter regime
-      camera.yPos = this.virtualJumpY; // camera.yPos = 0;
+    } else {// usually for fpshooter regime
+      // camera.yPos = this.virtualJumpY;
+      // camera.yPos = 0;
+      // leave it zero by default lets dirigent from top
     }
 
     camera.zPos -= Math.cos(degToRad(camera.yaw)) * camera.speed;